Support soft re-rolling
Adding paragraph specifying how to soft re-roll your character in case you are bad at math or certain skills are never being utilized. This fixes something that I find very annoying in nearly every game (pen and paper, as well as video) which is lacking flexibility when you build your character wrong. The most usual way to avoid this is to find builds online ahead of time and follow them religiously, which removes a lot of fun factor.